Only sending video streams for predecessors.  Also, reduced video quality.
I want to document here my anger at the following people:
 - whoever decided having Array.slice and Array.splice was a good idea
 - whoever thought that passing messages by reference to far-flung event
handlers was a good idea
 - whoever thought in-place mutation was a good idea.

One of my event handlers accidentally used splice instead of slice,
which instead of causing a local bug, changed the argument to the event
handler, so that a separate event handler in a completely different part
of the program did the wrong thing.  For all the world, it looked like
one client was receiving a different network message from all the
others.

Grumble grumble.